History/pushState support
Detect whether or not the browser supports the history API. If yes use
it by default (maybe some config could also change that). The code was
picked from `go-humble/router` thanks to him.

Unfortunately there's no event on `pushState` so listening to the
`popstate` event is not enough. For that reason I needed to manually
dispatch `popstate` on `GoWithParam` method. Not sure it's the best
way to handle that.